Update `obtainedMark` type from Float to Double across the app
Updated the `obtainedMark` field to use the `Double` type instead of `Float` for improved precision. Modified affected entities, DTOs, repositories, and queries accordingly. Also optimized and removed unused imports and parameters in related classes.

+++ src/main/java/com/edusyspro/api/dto/ScoreDTO.java
@@ -16,7 +16,7 @@ public class ScoreDTO {
     private Long id;
     private AssignmentDTO assignment;
     private StudentDTO student;
-    private Float obtainedMark;
+    private Double obtainedMark;
     private Boolean isPresent;
 
     public Score toEntity() {
